==> analyses/04-Univar_Bivar_Assoc.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# Univariate and Bivariate Summary Statistics
```{r, include=FALSE}
knitr::opts_chunk$set(echo = T, warning = F, message = F, eval = T, results = 'hide',
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = here::here())
```
```{r}
#------------------------------------------------------------------ 
# Purpose of this script is to do basic bivariate analyses
#------------------------------------------------------------------ 
source("R/00-functions_basic.R") 
source("R/00-functions_epi.R") 
library(tidyverse)
library(survey)
library(srvyr) #wrap the survey package in dplyr syntax
library(tableone)
set.seed(48)

#......................
# Import Data
#......................
dt <- readRDS("data/derived_data/vividepi_recode_completecases.rds")
dcdr <- readxl::read_excel(path = "model_datamaps/sub_DECODER_covariate_map_v3.xlsx", sheet = 1) %>% 
  dplyr::mutate(risk_factor_raw = ifelse(is.na(risk_factor_raw), "n", risk_factor_raw),
                risk_factor_model = ifelse(is.na(risk_factor_model), "n", risk_factor_model))

dtsrvy <- makecd2013survey(survey = dt)

ge <- readRDS(file = "data/raw_data/dhsdata/VivIDge.RDS")

```


## Prevalence
```{r}
popN.wi <- dtsrvy %>% 
  dplyr::mutate(cnt = 1) %>% 
  dplyr::summarise(
    n = srvyr::survey_total(cnt, vartype = c("se", "ci"), level = 0.95)
  ) %>% 
  dplyr::mutate_if(is.numeric, round, 2)


malN.wi <- dtsrvy %>% 
  dplyr::summarise(
    pv18sn = srvyr::survey_total(pv18s, vartype = c("se", "ci"), level = 0.95),
    pfldhn = srvyr::survey_total(pfldh, vartype = c("se", "ci"), level = 0.95),
  ) %>% 
  dplyr::mutate_if(is.numeric, round, 2)

```


Overall, we have `r nrow(dt)` individuals considered in the study which results in a weighted population of:   

```{r, results='asis'}
knitr::kable(popN.wi)
```



### National Prevalence
Fit basic intercept weighted-glm to derive weighted national prevalence. 

```{r, results='asis'}

#-------------------------------------------------------- 
# Basic Descriptive Statistics
#-------------------------------------------------------- 
# national prevalence
dtsrvy %>% 
  dplyr::mutate(count = 1) %>% 
  dplyr::summarise(n = srvyr::survey_total(count, vartype = c("se", "ci")),
                   pvn = srvyr::survey_total(pv18s, vartype = c("se", "ci")),
                   pfn = srvyr::survey_total(pfldh, vartype = c("se", "ci"))) %>% 
  knitr::kable(.)
```

```{r, results='asis'}
drcpv <- survey::svyglm(pv18s ~ 1, design = dtsrvy)
drcpv <- broom::tidy(drcpv, conf.int = T)

# national prevalence
drcpf <- survey::svyglm(pfldh ~ 1, design = dtsrvy)
drcpf <- broom::tidy(drcpf, conf.int = T)

rbind.data.frame( cbind.data.frame(species = "Pv", drcpv),
                  cbind.data.frame(species = "Pf", drcpf)
) %>% 
  dplyr::select(-c("term")) %>% 
  dplyr::rename(Prevalence = estimate) %>% 
  knitr::kable()


```

### Cluster Level Prevalence 
**N.B. Everyone is weighted the same in the cluster, which results in the same cluster level prevalences whether you consider weights or not**. However, this will results in  slightly different numerator (e.g. $n_numerator$) and denominators between the weighted and unweighted approach. Caution is warranted when reporting weighted versus unweighted Ns and Standard Errors at the cluster-level for this reason. 

NB Given the small Ns (<30) in many of the clusters, we cannot approximate the binomial distribution as a normal distribution, and thus our confidence interval calculations need to be performed with exact methods to avoid boundary issues (e.g. prevalences < 0 or > 100%). I used the binomial exact approach. 


```{r, results='asis', fig.align='center', fig.width=11, fig.height=8}
# cluster-level prevalence, because everyone is weighted the same in the cluster,round to use dhs weights
clst <- dtsrvy %>% 
  dplyr::mutate(count = 1) %>% 
  dplyr::group_by(hv001) %>% 
  dplyr::summarise(n = srvyr::survey_total(count, vartype = c("se", "ci")),
                   pv18sn = srvyr::survey_total(pv18s, vartype = c("se", "ci")),
                   pfldhn = srvyr::survey_total(pfldh, vartype = c("se", "ci"))) %>% 
  dplyr::mutate(
    pv18sbinomtest = purrr::map2(pv18sn, n, function(x,y)
    {binom.test(x = round(x), n = round(y))}),
    pv18sprev = pv18sn/n,
    pvconfint = purrr::map(pv18sbinomtest, "conf.int"),
    pv18sL95 = purrr::map_dbl(pvconfint, 1),
    pv18sU95 = purrr::map_dbl(pvconfint, 2),
    
    pfldhbinomtest = purrr::map2(pfldhn, n, function(x,y)
    {binom.test(x = round(x), n = round(y))}),
    pfldhprev = pfldhn/n,
    pfldhconfint = purrr::map(pfldhbinomtest, "conf.int"),
    pfldhL95 = purrr::map_dbl(pfldhconfint, 1),
    pfldhU95 = purrr::map_dbl(pfldhconfint, 2)
  )

# aesthetics for plot
genosf <- ge %>% 
  dplyr::mutate(longnum = sf::st_coordinates(geometry)[,1],
                latnum = sf::st_coordinates(geometry)[,2])
sf::st_geometry(genosf) <- NULL
#......................
# make cluster plot
#......................
clst.prev.plotObj <- clst %>% 
  dplyr::left_join(., y = genosf, by = "hv001") %>% 
  dplyr::select(c("adm1name", "hv001", "pv18sprev", "pv18sL95", "pv18sU95",
                  "pfldhprev", "pfldhL95", "pfldhU95")) %>% 
  tidyr::gather(., key = "estlevel", value = "est", 3:ncol(.)) %>% 
  dplyr::mutate(species = ifelse(grepl("pv", estlevel), "Pv", "Pf")) %>% 
  dplyr::select(c("adm1name", "hv001", "species", dplyr::everything())) %>% 
  dplyr::mutate(estlevel = gsub("pv18s", "", estlevel),
                estlevel= gsub("pfldh", "", estlevel)) %>% 
  tidyr::pivot_wider(data = ., names_from = "estlevel", values_from = "est") %>% 
  ggplot() + 
  geom_pointrange(aes(x = factor(hv001), y=prev, ymin=L95, ymax=U95, color = factor(species)), size =0.5, alpha = 0.5) +
  coord_flip() + 
  facet_wrap(~adm1name, scales = "free", ncol = 4) +
  scale_color_viridis_d("Species") + 
  ggtitle("Cluster-Level Prevelance of Pv and Pf Malaria") +
  vivid_theme + 
  theme(plot.title = element_text(family = "Helvetica", face = "bold", hjust = 0.5, size = 12),
        legend.position = "bottom",
        legend.title = element_text(family = "Helvetica", face = "bold", vjust = 0.5, size = 8),
        legend.text = element_text(family = "Helvetica", hjust = 0.5, vjust = 0.5, size = 7, angle = 0),
        strip.text = element_text(family = "Helvetica", hjust = 0.5, size = 6, face = "bold"),
        axis.title = element_blank(),
        axis.text.y = element_blank(),
        axis.text.x = element_text(family = "Helvetica", hjust = 0.5, size = 6, angle = 90),
        axis.ticks.y = element_blank()
  )


plot(clst.prev.plotObj)

dir.create(path = "results/figures/", recursive = T)
jpeg(filename = "results/figures/cluster_prevalence_pointranges.jpg",
     width = 11, height = 8, units = "in", res=500)
plot(clst.prev.plotObj)
graphics.off()


```

#### Cluster Level Prevalence and SEs
```{r, results='asis', fig.align='center', fig.width=11, fig.height=8}

# prov centroid for color
DRCprov <- readRDS("data/map_bases/vivid_DRCprov.rds")
cntr <- sf::st_centroid(DRCprov) %>% 
  dplyr::mutate(longnum = sf::st_coordinates(geometry)[,1])
# decoder for cluster to admin
admn_hv001_dcdr <- genosf %>% 
  dplyr::select(c("adm1name", "hv001"))

# get plot
clst.prev.plotObj <- clst %>% 
  dplyr::left_join(., y = admn_hv001_dcdr, by = "hv001") %>% 
  dplyr::left_join(., cntr, by = "adm1name") %>% 
  dplyr::select(c("adm1name", "hv001", "longnum", "n", "pv18sn", "pv18sprev", "pv18sL95", "pv18sU95")) %>%
  #dplyr::filter(pv18sprev != 0) %>% 
  dplyr::mutate(hv001 = factor(hv001),
                adm1name = forcats::fct_reorder(adm1name, longnum, min),
                adm1namenum = as.numeric(adm1name),
                hv001 = forcats::fct_reorder(hv001, adm1namenum, min)) %>% 
  ggplot() + 
  geom_pointrange(aes(x = hv001, y=pv18sprev, ymin=pv18sL95, ymax=pv18sU95,
                      color = adm1name), 
                  size =0.5, alpha = 0.5) +
  scale_color_viridis_d("Province") + 
  ggtitle("Cluster-Level Prevelance of Pv Malaria") +
  theme_bw() + 
  theme(plot.title = element_text(family = "Helvetica", face = "bold", hjust = 0.5, size = 12),
        legend.position = "bottom",
        legend.title = element_text(family = "Helvetica", face = "bold", vjust = 0.5, size = 8),
        legend.text = element_text(family = "Helvetica", hjust = 0.5, vjust = 0.5, size = 7, angle = 0),
        axis.title = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_text(family = "Helvetica", hjust = 0.5, size = 6, angle = 90),
        axis.ticks.x = element_blank()
  )


plot(clst.prev.plotObj)

dir.create(path = "results/figures/", recursive = T)
jpeg(filename = "results/figures/pos_cluster_prevalence_pointranges_E-W_colorgrad.jpg",
     width = 11, height = 8, units = "in", res=500)
plot(clst.prev.plotObj)
graphics.off()

```

### Cluster Level Prevalences Table (Unweighted Ns)
```{r, results='asis'}

clust_no_weights <- dt %>%
  group_by(hv001) %>% 
  dplyr::summarise(
    n = dplyr::n(),
    pv18sn = sum(pv18s),
    pv18sprev = pv18sn/n
  ) 

clust_no_weights %>% 
  dplyr::mutate_if(is.numeric, round, 4) %>% 
  DT::datatable(., extensions='Buttons',
                options = list(
                  searching = T,
                  pageLength = 5,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))

```

Not considering weights, among the `r nrow(clst)` clusters considered, `r sum(clust_no_weights$pv18sn >= 1)` clusters had at least one infection and `r sum(clust_no_weights$pv18sn == 1)` had exactly one infection. 

### Cluster Level Prevalences Table (Weighted Ns)
```{r, results='asis'}

clst %>% 
  dplyr::select(c("hv001", "n", "pv18sn", "pfldhn",
                  "pv18sprev", "pv18sL95", "pv18sU95",
                  "pfldhprev", "pfldhL95", "pfldhU95")) %>% 
  dplyr::mutate_if(is.numeric, round, 4) %>% 
  DT::datatable(., extensions='Buttons',
                options = list(
                  searching = T,
                  pageLength = 5,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))

```




## Duffy Negativity
### Counts and Prevalences
```{r}
# three samples with duffy negativity
duffypos <- tolower(c("C1T5C", "G9M2E", "J3Q9M"))

duff <- dt %>% 
  dplyr::mutate(duffy = ifelse(hivrecode_barcode %in% duffypos, 1, 0)) %>% 
  dplyr::summarise(
    duff_pos = sum(duffy),
    duff_screen = sum(pv18s)) %>%  
  dplyr::mutate(
    duffbinomtest = purrr::map2(duff_pos, duff_screen, function(x,y)
    {binom.test(x = x, n = y)}),
    duffprev = duff_pos/duff_screen,
    duffconfint = purrr::map(duffbinomtest, "conf.int"),
    duffL95 = purrr::map_dbl(duffconfint, 1),
    duffU95 = purrr::map_dbl(duffconfint, 2))


```

Note, prevalences below are not in percentages
```{r, results='asis'}

duff %>% 
  dplyr::select(c("duff_pos", "duff_screen", "duffprev", "duffL95", "duffU95")) %>% 
  dplyr::mutate_if(is.numeric, round, 4) %>% 
  knitr::kable(.)

```


## Coinfections
### Unweighted Counts
```{r, results='asis'}

# coinfections unweighted
coinfxn <- dt %>% 
  dplyr::mutate(pfmono =  ifelse(pfldh == 1 & pv18s == 0, 1, 0),
                pvmono =  ifelse(pfldh == 0 & pv18s == 1, 1, 0),
                pfpv =  ifelse(pfldh == 1 & pv18s == 1, 1, 0),
  ) %>% 
  dplyr::summarise(
    "pf" = sum(pfmono),
    "pv" = sum(pvmono),
    "pf/pv" = sum(pfpv)
  )

knitr::kable(coinfxn)



```

### Weighted Counts
```{r, results='asis'}
# need weighed counts
coinfxn.wi <- dtsrvy %>% 
  dplyr::mutate(pfmono =  ifelse(pfldh == 1 & pv18s == 0,  1, 0),
                pvmono =  ifelse(pfldh == 0 & pv18s == 1,  1, 0),
                pfpv =  ifelse(pfldh == 1 & pv18s == 1,  1, 0),
  ) %>% 
  dplyr::summarise(
    "pf" = srvyr::survey_total(x=pfmono, vartype = c("ci"), level = 0.95),
    "pv" = srvyr::survey_total(x=pvmono, vartype = c("ci"), level = 0.95),
    "pf/pv" = srvyr::survey_total(x=pfpv, vartype = c("ci"), level = 0.95),
  )

coinfxn.wi %>% 
  dplyr::select(-c(dplyr::ends_with("_low"), dplyr::ends_with("_upp"))) %>% 
  knitr::kable()

coinfxn.wi %>% 
  dplyr::select(c(dplyr::ends_with("_low"), dplyr::ends_with("_upp"))) %>% 
  knitr::kable()

```

# Table One(s)
## _P. vivax_ Infected vs. Uninfected
```{r, results='asis'}
#------------------------------------------------------------------------------------- 
# TABLE ONE -- bivariate analyses, descriptive
#------------------------------------------------------------------------------------- 

#.......................
# Pvivax 
#.......................
pvivrskfctr <- dcdr$column_name[dcdr$risk_factor_raw == "y"]

pvivtbl1 <- tableone::svyCreateTableOne(data = dtsrvy,
                                        strata = "pv18s_fctb",
                                        vars = pvivrskfctr,
                                        includeNA = T,
                                        test = F)

tableone::kableone(pvivtbl1)

```

## _P. falciparum_ Infected vs. Uninfected
```{r, results='asis'}

#.......................
# Pfalciparum 
#.......................
pfalrskfctr <- dcdr$column_name[dcdr$risk_factor_raw == "y" & dcdr$column_name != "pfldh_fctb" ]
pfalrskfctr <- c("pv18s_fctb", pfalrskfctr)

pfaltbl1 <- tableone::svyCreateTableOne(
  data = dtsrvy,
  strata = "pfldh_fctb",
  vars = pfalrskfctr,
  includeNA = T,
  test = F)

tableone::kableone(pfaltbl1)

```

## Infected with Malaria vs. Not-Infected
```{r, results='asis'}
#------------------------------------------------------------------------------------
# Make Pv Cases, Pf Cases, and Non-Cases Table
#------------------------------------------------------------------------------------
dt.cases <- dt %>% 
  dplyr::mutate(
    case = ifelse(( pv18s == 1 | pfldh ==1 ), 1, 0),
    case_fctb = factor(case, levels = c(0,1), labels = c("noncase", "case"))
  )

dt.cases.srvy <- makecd2013survey(dt.cases)
casesrskfctr <- dcdr$column_name[dcdr$risk_factor_raw == "y"]
casestbl1 <- tableone::svyCreateTableOne(
  data = dt.cases.srvy,
  strata = "case_fctb",
  vars = casesrskfctr,
  includeNA = T,
  test = F)

tableone::kableone(casestbl1)

```

# Table Two(s)
N.B. We have no scaled (i.e. mean-centered and standardized) all continous variables with the exception of the wlth variable, as it was already a factor score. Dichotomized variables have the presumed protective level as the reference level.   

N.B. Parametric association test using GEE (`svyglm` is performing GEE with the sampling weights considered "repeat" observations). Given that we are not considering confounding in these estimates, they are purely assocations and should not be overly interpreted. 

```{r}
#-----------------------------------------------------------------------------------------
# TABLE 2
# Parametric, Bivariate Analysis
# note, that svyglm is really doing GEE
#----------------------------------------------------------------------------------------

# note overwriting risk factors to have scaled models now
#.......................
# Pvivax 
#.......................
pvivrskfctr <- dcdr$column_name[dcdr$risk_factor_model == "y"]
pvivrskfctr_models <- data.frame(outcome = rep("pv18s", length(pvivrskfctr)), 
                                 covar = pvivrskfctr, stringsAsFactors=FALSE)

pvivrskfctr_models$glmlogit <- purrr::pmap(pvivrskfctr_models, .f=fitsvyglmlogit)
pvivrskfctr_models$glmlogit_tidy <- purrr::map(pvivrskfctr_models$glmlogit,
                                               .f=function(x){
                                                 broom::tidy(x, exponentiate=TRUE, conf.int=TRUE)}
)
pvivrskfctr_est <- pvivrskfctr_models$glmlogit_tidy %>% 
  bind_rows() %>% filter(term != "(Intercept)") %>% 
  mutate_if(is.numeric, round, 2)


```


```{r}
#.......................
# Pfalciparum 
#.......................
pfalrskfctr <- dcdr$column_name[dcdr$risk_factor_model == "y" & dcdr$column_name != "pfldh_fctb"]
pfalrskfctr <- c("pv18s_fctb", pfalrskfctr)

pfalrskfctr_models <- data.frame(outcome = rep("pfldh", length(pfalrskfctr)), 
                                 covar = pfalrskfctr, stringsAsFactors=FALSE)

pfalrskfctr_models$glmlogit <- purrr::pmap(pfalrskfctr_models, .f=fitsvyglmlogit)


pfalrskfctr_models$glmlogit_tidy <- purrr::map(pfalrskfctr_models$glmlogit,
                                               .f=function(x){
                                                 broom::tidy(x, exponentiate=TRUE, conf.int=TRUE)})


pfalrskfctr_est <- pfalrskfctr_models$glmlogit_tidy %>% 
  bind_rows() %>% filter(term != "(Intercept)") %>% 
  mutate_if(is.numeric, round, 2)



```

```{r}
#------------------------------------------------------------------
# Combine Table 1 and 2
#------------------------------------------------------------------ 

#.......................
# Pvivax
#.......................
pvivtbl1df <- tableone2dataframe(pvivtbl1, columnnames = c("Covariates",
                                                           "Pvivax-Negative",
                                                           "Pvivax-Positive",
                                                           "matchcol"))
pvivtbl1df <- dcdr %>% 
  dplyr::select(c("column_name", "var_label")) %>% 
  dplyr::rename(matchcol = column_name) %>% 
  dplyr::left_join(pvivtbl1df, ., by = "matchcol") %>% 
  dplyr::select(c("var_label", dplyr::everything()))

# manual fix
pvivtbl1df$matchcol[pvivtbl1df$matchcol == "wtrdist_cont_clst"] <- "wtrdist_cont_log_scale_clst"



pvivriskfactortable <- mergetableone2table(tableonedf = pvivtbl1df,
                                           tabletwoestdf = pvivrskfctr_est)

pvivriskfactortable <- pvivriskfactortable %>% 
  dplyr::select(-c("matchcol"))

#.......................
# Pfalciparum 
#.......................
pfaltbl1df <- tableone2dataframe(pfaltbl1, columnnames = c("Covariates",
                                                           "Pfalciparum-Negative",
                                                           "Pfalciparum-Positive",
                                                           "matchcol"))

pfaltbl1df <- dcdr %>% 
  dplyr::select(c("column_name", "var_label")) %>% 
  dplyr::rename(matchcol = column_name) %>% 
  dplyr::left_join(pfaltbl1df, ., by = "matchcol") %>% 
  dplyr::select(c("var_label", dplyr::everything()))

# manual fix
pfaltbl1df$matchcol[pfaltbl1df$matchcol == "wtrdist_cont_clst"] <- "wtrdist_cont_log_scale_clst"


pfalriskfactortable <- mergetableone2table(tableonedf = pfaltbl1df,
                                           tabletwoestdf = pfalrskfctr_est)


pfalriskfactortable <- pfalriskfactortable %>% 
  dplyr::select(-c("matchcol"))



#.......................
# Cases to DF 
#.......................
casestbl1df <- tableone2dataframe(casestbl1, columnnames = c("Covariates",
                                                             "Case-Negative",
                                                             "Case-Positive",
                                                             "matchcol"))

casestbl1df <- dcdr %>% 
  dplyr::select(c("column_name", "var_label")) %>% 
  dplyr::rename(matchcol = column_name) %>% 
  dplyr::left_join(casestbl1df, ., by = "matchcol") %>% 
  dplyr::select(c("var_label", dplyr::everything()))




```

## _P. vivax_ Risk Bivariate **Associations**
```{r, results='asis'}

DT::datatable(pvivriskfactortable, 
              rownames = F,
              options = list(
                searching = F,
                pageLength = 30,
                dom = 'Bfrtip', 
                buttons = c('csv')),
              caption = htmltools::tags$caption(
                style = 'caption-side: bottom; text-align: center;',
                "P vivax Risk Factor Table with Counts and Prevalence Odds Ratio Bivariate Associations"))

```

#### _P. vivax_ Bed Nets

```{r}
dt_sens <- readRDS("data/derived_data/vividepi_recode.rds")

dcdr_sens <- readxl::read_excel(path = "model_datamaps/sub_DECODER_covariate_map_v3.xlsx", sheet = 1) %>% 
  dplyr::pull(c("column_name"))
sf::st_geometry(dt_sens) <- NULL
dt_sens <- dt_sens  %>% 
  dplyr::select(c("pv18s", "pfldh", "po18s", dcdr_sens, "hml20_fctb")) %>% 
  dplyr::filter(complete.cases(.)) %>% 
  dplyr::mutate(hml20_fctb = relevel(hml20_fctb, ref = "yes"))
dtsrvy_sens <- makecd2013survey(survey = dt_sens)


basicnet <- survey::svyglm("pv18s ~ hml20_fctb",
                           design = dtsrvy_sens,
                           family = quasibinomial(link="logit")) 
```

##### LLIN Def

```{r, results='asis'}
knitr::kable(broom::tidy(basicnet, exponentiate=TRUE, conf.int=TRUE))
```

##### ITN Def

```{r}
longnet <- survey::svyglm("pv18s ~ ITN_fctb",
                          design = dtsrvy_sens,
                          family = quasibinomial(link="logit"))
```

```{r, results='asis'}
knitr::kable(broom::tidy(longnet, exponentiate=TRUE, conf.int=TRUE))
```

## _P. falciparum_ Risk Bivariate **Associations**
```{r, results='asis'}
DT::datatable(pfalriskfactortable, 
              rownames = F,
              options = list(
                searching = F,
                pageLength = 30,
                dom = 'Bfrtip', 
                buttons = c('csv')),
              caption = htmltools::tags$caption(
                style = 'caption-side: bottom; text-align: center;',
                "P falciparum Risk Factor Table with Counts and Prevalence Odds Ratio Bivariate Associations"))

```


## Malaria Risk Association Figure

```{r, results='asis', fig.align='center', fig.width=8, fig.height=8}
#...................................................
# Tidy
#...................................................
pvriskest <- pvivrskfctr_models$glmlogit_tidy %>% 
  bind_rows() %>% filter(term != "(Intercept)") %>% 
  mutate_if(is.numeric, round, 2) %>% 
  dplyr::rename(column_name = term) %>% 
  dplyr::mutate(
    column_name = ifelse(grepl("_fctb_clst", column_name),
                         stringr::str_extract(column_name, "[ -~]+_fctb_clst"),
                         column_name),
    column_name = ifelse(!grepl("_fctb_clst", column_name) & grepl("_fctb", column_name),
                         stringr::str_extract(column_name, "[ -~]+_fctb"),
                         column_name)
  )

pvriskest <- dplyr::left_join(pvriskest, dcdr, by = "column_name")
orderrf <- pvriskest %>% 
  dplyr::arrange(level) %>% 
  dplyr::select(abridged_var_label) %>% 
  unlist(.)

# PF
pfriskest <- pfalrskfctr_models$glmlogit_tidy %>% 
  bind_rows() %>% filter(term != "(Intercept)") %>% 
  mutate_if(is.numeric, round, 2) %>% 
  dplyr::rename(column_name = term)  %>% 
  dplyr::mutate(
    column_name = ifelse(grepl("_fctb_clst", column_name),
                         stringr::str_extract(column_name, "[ -~]+_fctb_clst"),
                         column_name),
    column_name = ifelse(!grepl("_fctb_clst", column_name) & grepl("_fctb", column_name),
                         stringr::str_extract(column_name, "[ -~]+_fctb"),
                         column_name)
  )

pfriskest <- dplyr::left_join(pfriskest, dcdr, by = "column_name")
orderrf <- pfriskest %>% 
  dplyr::arrange(level) %>% 
  dplyr::select(abridged_var_label) %>% 
  unlist(.)

#...................................................
# Make Plots
#...................................................

pv_bivar_rf_plot <- pvriskest %>% 
  dplyr::mutate(abridged_var_label = factor(abridged_var_label, levels = orderrf, ordered = T)) %>% 
  ggplot() +
  geom_hline(yintercept = 1, color = "#cb181d", linetype = "dashed") +
  geom_pointrange(aes(x = abridged_var_label, y = estimate, 
                      ymin = conf.low, ymax = conf.high,
                      color = factor(level))) +
  scale_color_manual("Level", values = c("#0868ac", "#4eb3d3")) +
  coord_flip() + 
  ggtitle(expression(bolditalic("P. vivax"))) +
  ylab("Prevalence Odds Ratio") + 
  theme(
    plot.title =  element_text(family = "Helvetica", face = "bold", vjust = 0.5, hjust = 0.5, size = 14),
    axis.title.x = element_text(family = "Helvetica", face = "bold", vjust = 0.5, hjust = 0.5, size = 12, angle = 45),
    axis.text = element_text(family = "Helvetica", vjust = 0.5, hjust = 0.5, size = 11),
    axis.title.y = element_blank(),
    legend.title = element_text(family = "Helvetica", face = "bold", vjust = 0.5, hjust = 0.5, size = 12),
    legend.text = element_text(family = "Helvetica", vjust = 0.5, hjust = 0.5, size = 10, angle = 0),
    legend.position = "right",
    axis.line.x = element_line(color = "black", size = 1.5),
    axis.line.y = element_blank(),
    axis.ticks.y = element_blank(),
    panel.background = element_rect(fill = "transparent"),
    plot.background = element_rect(fill = "transparent"),
    panel.grid = element_blank(),
    panel.border = element_blank())


pf_bivar_rf_plot <- pfriskest %>% 
  dplyr::mutate(abridged_var_label = factor(abridged_var_label, levels = orderrf, ordered = T)) %>%
  ggplot() +
  geom_hline(yintercept = 1, color = "#cb181d", linetype = "dashed") +
  geom_pointrange(aes(x = abridged_var_label, y = estimate, 
                      ymin = conf.low, ymax = conf.high,
                      color = factor(level))) +
  scale_color_manual("Level", values = c("#006d2c", "#41ae76")) +
  ylim(c(0,2)) + 
  coord_flip() + 
  ggtitle(expression(bolditalic("P. falciparum"))) +
  ylab("Prevalence Odds Ratio") + 
  theme(
    plot.title =  element_text(family = "Helvetica", face = "bold", vjust = 0.5, hjust = 0.5, size = 14),
    axis.title.x = element_text(family = "Helvetica", face = "bold", vjust = 0.5, hjust = 0.5, size = 12, angle = 45),
    axis.text = element_text(family = "Helvetica", vjust = 0.5, hjust = 0.5, size = 11),
    axis.title.y = element_blank(),
    legend.title = element_text(family = "Helvetica", face = "bold", vjust = 0.5, hjust = 0.5, size = 12),
    legend.text = element_text(family = "Helvetica", vjust = 0.5, hjust = 0.5, size = 10, angle = 0),
    legend.position = "right",
    axis.line.x = element_line(color = "black", size = 1.5),
    axis.line.y = element_blank(),
    axis.ticks.y = element_blank(),
    panel.background = element_rect(fill = "transparent"),
    plot.background = element_rect(fill = "transparent"),
    panel.grid = element_blank(),
    panel.border = element_blank())

cowplot::plot_grid(pv_bivar_rf_plot, 
                   pf_bivar_rf_plot, 
                   nrow = 1, 
                   align = "h",
                   labels = c("A", "B"))


# svglite::svglite(file = "results/figures/Figure1B.svg")
jpeg("results/figures/PfPv_Bivariate_dotplotRR.jpg", 
     width = 11, height = 8, units = "in", res = 500)
cowplot::plot_grid(pv_bivar_rf_plot, 
                   pf_bivar_rf_plot, 
                   nrow = 1, 
                   align = "h",
                   labels = c("A", "B"))
graphics.off()


```

```{r}

#-----------------------------------------------------------------------------
# housekeeping -- Save out
#-----------------------------------------------------------------------------
save(pvivtbl1df, pfaltbl1df, # table one output 
     pvivrskfctr_models, pfalrskfctr_models, # model datatframes
     pvivriskfactortable, pfalriskfactortable, # final out table for report
     casestbl1df, # for prettier table 1
     file = "results/bivariate_model_results.rda")


#-----------------------------------------------------------------------------
# housekeeping -- write out
#-----------------------------------------------------------------------------
# Case/NonCase Dist Tables
dir.create(path = "results/tables/", recursive = T)
readr::write_csv(x = pvivtbl1df, 
                 path = "results/tables/pvivtblones_counts.csv",
                 na = "")

readr::write_csv(x = pfaltbl1df, 
                 path = "results/tables/pfaltblones_counts.csv",
                 na = "")

readr::write_csv(x = casestbl1df, 
                 path = "results/tables/cases_noncases_tblones_counts.csv",
                 na = "")





readr::write_csv(x = pvivriskfactortable, 
                 path = "results/tables/pvivriskfactortable_tabletwos.csv",
                 na = "")

readr::write_csv(x = pfalriskfactortable, 
                 path = "results/tables/pfalriskfactortable_tabletwos.csv",
                 na = "")


```
